Update conditions to reflect degraded and resyncing
The condition updates are when processing as secondary

- Resyncing
  - Is false on any errors
  - Is false if degraded is false
  - Is true, if resync is successful and volume is not
    yet "ready" to use
- Degraded
  - Is true on errors
  - Is true UNTIL resync returns "ready" response
  - Is false ONLY when resync returns "ready"

As Degraded moves to false, resyncing will move to
false as well, other than cases where there are
errors, or resync status is not yet determined. Resync
status is, for example, undetermined if we issued a
demote volume request, from markVolAsSecondary.

Resyncing: denotes if resync is active
Degraded: denotes local secondary contents are not
in full sync with remote, and that happens only when
remote is marked Secondary and local is also Secondary
and the final sync is complete.

Hence, Resyncing is false when degraded is false, as
there is no more data to sync.

Completed is always the initial condition to check before
looking at other conditions, to ensure operation was a
success. IOW, Completed being false, currently negates
any value in the other 2 conditions, and should typically
be Resync: false and Degraded: true.

Fixes: #101

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit d33dbd5432714a04f69da59682cf25d2a20c911d)